requirement.txt

Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
Flask-Session==0.8.0
python-dotenv==1.0.1